@@ -3,15 +3,16 @@ from __future__ import annotations
 
 import argparse
 import json
+import re
+import subprocess
 import sys
 from pathlib import Path
-from typing import Any, Dict, List
-import re
+from typing import Any
 
 try:
-    import yaml  # type: ignore
+    import yaml  # type: ignore[import-untyped,unused-ignore]
 except Exception:  # pragma: no cover
-    yaml = None
+    yaml = None  # type: ignore[assignment]
 
 FRONT_MATTER_DELIM = re.compile(r"^---\s*$")
 
@@ -29,24 +30,28 @@ def read_text(p: Path) -> str:
     return p.read_text(encoding="utf-8")
 
 
-def extract_front_matter(md_text: str) -> Dict[str, Any]:
+def extract_front_matter(md_text: str) -> dict[str, Any]:
     lines = md_text.splitlines()
     if not lines or not FRONT_MATTER_DELIM.match(lines[0]):
-        raise ValueError("Missing starting '---' for front matter")
+        msg = "Missing starting '---' for front matter"
+        raise ValueError(msg)
     end_idx = None
     for i in range(1, len(lines)):
         if FRONT_MATTER_DELIM.match(lines[i]):
             end_idx = i
             break
     if end_idx is None:
-        raise ValueError("Missing closing '---' for front matter")
+        msg = "Missing closing '---' for front matter"
+        raise ValueError(msg)
 
     fm_text = "\n".join(lines[1:end_idx])
     if yaml is None:
-        raise RuntimeError("PyYAML not installed. Please add 'pyyaml'.")
+        msg = "PyYAML not installed. Please add 'pyyaml'."
+        raise RuntimeError(msg)
     meta = yaml.safe_load(fm_text) or {}
     if not isinstance(meta, dict):
-        raise ValueError("Front matter must be a YAML mapping")
+        msg = "Front matter must be a YAML mapping"
+        raise ValueError(msg)
     return meta
 
 
@@ -58,6 +63,11 @@ def main() -> int:
         type=Path,
         default=None,
         help="Output path (default: <root>/index/skills-index.json)",
+    )
+    ap.add_argument(
+        "--with-embeddings",
+        action="store_true",
+        help="Also rebuild embeddings after building index",
     )
     args = ap.parse_args()
 
@@ -71,7 +81,7 @@ def main() -> int:
         return 2
     index_dir.mkdir(parents=True, exist_ok=True)
 
-    entries: List[Dict[str, Any]] = []
+    entries: list[dict[str, Any]] = []
     for skill_md in sorted(skills_dir.glob("*/SKILL.md")):
         meta = extract_front_matter(read_text(skill_md))
         missing = [k for k in META_FIELDS if k not in meta]
@@ -100,6 +110,20 @@ def main() -> int:
     out.parent.mkdir(parents=True, exist_ok=True)
     out.write_text(json.dumps(entries, indent=2) + "\n", encoding="utf-8")
     print(f"Wrote {out} with {len(entries)} entr(y/ies)")
+
+    # Optionally rebuild embeddings
+    if args.with_embeddings:
+        print("\nRebuilding embeddings...")
+        embeddings_script = Path(__file__).parent / "build_embeddings.py"
+        # S603: Subprocess call is safe - controlled script path and executable
+        result = subprocess.run(  # noqa: S603
+            [sys.executable, str(embeddings_script)], cwd=root, check=False
+        )
+        if result.returncode != 0:
+            msg = "Failed to build embeddings"
+            print(f"ERROR: {msg}", file=sys.stderr)
+            return result.returncode
+
     return 0
 
 
