@@ -19,11 +19,10 @@ from __future__ import annotations
 
 import argparse
 import sys
+from collections.abc import Iterable
 from pathlib import Path
-from typing import Iterable, List
 
-
-DEFAULT_DIRS: List[str] = [
+DEFAULT_DIRS: list[str] = [
     "index",
     "index/embeddings",
     "skills",
@@ -33,7 +32,7 @@ DEFAULT_DIRS: List[str] = [
     ".github/workflows",
 ]
 
-SKILL_SUBDIRS: List[str] = [
+SKILL_SUBDIRS: list[str] = [
     "examples",
     "resources",
     "scripts",
@@ -45,16 +44,19 @@ def _safe_join(root: Path, rel: str) -> Path:
     Join a relative path safely under root, preventing escape via '..' or absolute paths.
     """
     if rel.strip() == "":
-        raise ValueError("Empty relative path is not allowed.")
+        msg = "Empty relative path is not allowed."
+        raise ValueError(msg)
     p = Path(rel)
     if p.is_absolute():
-        raise ValueError(f"Absolute paths are not allowed: {rel}")
+        msg = f"Absolute paths are not allowed: {rel}"
+        raise ValueError(msg)
     # Normalize and ensure it's within root
     target = (root / p).resolve()
     root_resolved = root.resolve()
     if root_resolved not in target.parents and target != root_resolved:
         # If target isn't the root or inside it, reject.
-        raise ValueError(f"Unsafe path outside repo root: {rel}")
+        msg = f"Unsafe path outside repo root: {rel}"
+        raise ValueError(msg)
     return target
 
 
@@ -78,9 +80,11 @@ def _create_skill_skeleton(root: Path, slug: str, dry_run: bool) -> None:
     Only structure; does not create SKILL.md (separate generator handles that).
     """
     if not slug or slug.strip() == "":
-        raise ValueError("Skill slug cannot be empty.")
+        msg = "Skill slug cannot be empty."
+        raise ValueError(msg)
     if any(c in slug for c in r"\/:<>|*?\""):
-        raise ValueError(f"Skill slug contains invalid characters: {slug}")
+        msg = f"Skill slug contains invalid characters: {slug}"
+        raise ValueError(msg)
 
     base = _safe_join(root, f"skills/{slug}")
     _ensure_dir(base, dry_run)
@@ -88,7 +92,9 @@ def _create_skill_skeleton(root: Path, slug: str, dry_run: bool) -> None:
         _ensure_dir(base / sub, dry_run)
 
 
-def _create_structure(root: Path, dirs: Iterable[str], skills: Iterable[str], dry_run: bool) -> None:
+def _create_structure(
+    root: Path, dirs: Iterable[str], skills: Iterable[str], dry_run: bool
+) -> None:
     for d in dirs:
         target = _safe_join(root, d)
         _ensure_dir(target, dry_run)
@@ -99,7 +105,7 @@ def _create_structure(root: Path, dirs: Iterable[str], skills: Iterable[str], dr
             _create_skill_skeleton(root, normalized, dry_run)
 
 
-def parse_args(argv: List[str]) -> argparse.Namespace:
+def parse_args(argv: list[str]) -> argparse.Namespace:
     parser = argparse.ArgumentParser(
         description="Initialize Skills repository folder structure with .gitkeep files."
     )
@@ -133,7 +139,7 @@ def parse_args(argv: List[str]) -> argparse.Namespace:
     return parser.parse_args(argv)
 
 
-def main(argv: List[str]) -> int:
+def main(argv: list[str]) -> int:
     args = parse_args(argv)
 
     try:
@@ -152,7 +158,7 @@ def main(argv: List[str]) -> int:
                 print(f"ERROR: Could not create repo root: {exc}", file=sys.stderr)
                 return 2
 
-    dirs: List[str] = []
+    dirs: list[str] = []
     if not args.no_defaults:
         dirs.extend(DEFAULT_DIRS)
 
@@ -161,10 +167,10 @@ def main(argv: List[str]) -> int:
         if ed not in dirs:
             dirs.append(ed)
 
-    skills_list: List[str] = [s.strip() for s in args.skills.split(",") if s.strip()]
+    skills_list: list[str] = [s.strip() for s in args.skills.split(",") if s.strip()]
 
     # Always ensure top-level root gets a .gitkeep if requested structure is empty
-    # (We won’t create .gitkeep directly under root to avoid clutter; only inside directories.)
+    # (We will not create .gitkeep directly under root to avoid clutter; only inside directories.)
 
     try:
         _create_structure(root, dirs, skills_list, args.dry_run)
