@@ -6,12 +6,12 @@ import re
 import sys
 from dataclasses import dataclass
 from pathlib import Path
-from typing import Dict, List, Optional, Tuple
+from typing import Any
 
 try:
-    import yaml  # type: ignore
+    import yaml  # type: ignore[import-untyped,unused-ignore]
 except Exception:  # pragma: no cover
-    yaml = None
+    yaml = None  # type: ignore[assignment]
 
 FRONT_MATTER_DELIM = re.compile(r"^---\s*$")
 REQ_META_KEYS = {
@@ -49,14 +49,16 @@ SECRET_PATTERNS = [
     re.compile(r"(?i)secret\s*[:=]\s*[^\s]{6,}"),
 ]
 
+
 @dataclass
 class SkillValidationIssue:
     path: Path
     message: str
 
+
 @dataclass
 class FrontMatter:
-    meta: Dict
+    meta: dict[str, Any]
     body: str
 
 
@@ -68,7 +70,8 @@ def extract_front_matter(md_text: str) -> FrontMatter:
     """Extract YAML front matter delimited by '---' at the top of the file."""
     lines = md_text.splitlines()
     if not lines or not FRONT_MATTER_DELIM.match(lines[0]):
-        raise ValueError("Missing starting '---' for front matter")
+        msg = "Missing starting '---' for front matter"
+        raise ValueError(msg)
     # find closing '---'
     end_idx = None
     for i in range(1, len(lines)):
@@ -76,32 +79,34 @@ def extract_front_matter(md_text: str) -> FrontMatter:
             end_idx = i
             break
     if end_idx is None:
-        raise ValueError("Missing closing '---' for front matter")
+        msg = "Missing closing '---' for front matter"
+        raise ValueError(msg)
 
     fm_text = "\n".join(lines[1:end_idx])
     body = "\n".join(lines[end_idx + 1 :])
 
     if yaml is None:
-        raise RuntimeError(
-            "PyYAML not installed. Please add 'pyyaml' and re-run validator."
-        )
+        msg = "PyYAML not installed. Please add 'pyyaml' and re-run validator."
+        raise RuntimeError(msg)
     try:
         meta = yaml.safe_load(fm_text) or {}
     except Exception as e:  # pragma: no cover
-        raise ValueError(f"Failed to parse front matter YAML: {e}")
+        msg = f"Failed to parse front matter YAML: {e}"
+        raise ValueError(msg) from e
 
     if not isinstance(meta, dict):
-        raise ValueError("Front matter must be a YAML mapping (object)")
+        msg = "Front matter must be a YAML mapping (object)"
+        raise ValueError(msg)
 
     return FrontMatter(meta=meta, body=body)
 
 
-def find_code_blocks(text: str) -> List[Tuple[int, int]]:
+def find_code_blocks(text: str) -> list[tuple[int, int]]:
     """Return list of (start_line_idx, end_line_idx) for fenced code blocks."""
-    blocks: List[Tuple[int, int]] = []
+    blocks: list[tuple[int, int]] = []
     fence = re.compile(r"^```")
     lines = text.splitlines()
-    open_idx: Optional[int] = None
+    open_idx: int | None = None
     for i, ln in enumerate(lines):
         if fence.match(ln):
             if open_idx is None:
@@ -112,7 +117,7 @@ def find_code_blocks(text: str) -> List[Tuple[int, int]]:
     return blocks
 
 
-def first_examples_block_len(text: str) -> Optional[int]:
+def first_examples_block_len(text: str) -> int | None:
     """Return line count of the first code block under '## Examples'."""
     parts = re.split(r"^## Examples\s*$", text, flags=re.M)
     if len(parts) < 2:
@@ -127,10 +132,12 @@ def first_examples_block_len(text: str) -> Optional[int]:
 
 def has_token_budgets(text: str) -> bool:
     # simple check for presence of T1/T2/T3 in the Quality Gates or anywhere
-    return bool(re.search(r"\bT1\b", text) and re.search(r"\bT2\b", text) and re.search(r"\bT3\b", text))
+    return bool(
+        re.search(r"\bT1\b", text) and re.search(r"\bT2\b", text) and re.search(r"\bT3\b", text)
+    )
 
 
-def scan_secrets(text: str) -> Optional[str]:
+def scan_secrets(text: str) -> str | None:
     for pat in SECRET_PATTERNS:
         m = pat.search(text)
         if m:
@@ -138,8 +145,8 @@ def scan_secrets(text: str) -> Optional[str]:
     return None
 
 
-def validate_skill_file(path: Path) -> List[SkillValidationIssue]:
-    issues: List[SkillValidationIssue] = []
+def validate_skill_file(path: Path) -> list[SkillValidationIssue]:
+    issues: list[SkillValidationIssue] = []
     try:
         fm = extract_front_matter(read_text(path))
     except Exception as e:
@@ -192,7 +199,7 @@ def validate_skill_file(path: Path) -> List[SkillValidationIssue]:
             )
 
     # Oversized code blocks
-    for (start, end) in find_code_blocks(body):
+    for start, end in find_code_blocks(body):
         n = max(0, end - start - 1)
         if n > MAX_CODEBLOCK_LINES:
             issues.append(
@@ -231,7 +238,7 @@ def main() -> int:
         print("No skills found.")
         return 0
 
-    total_issues: List[SkillValidationIssue] = []
+    total_issues: list[SkillValidationIssue] = []
     for p in md_files:
         issues = validate_skill_file(p)
         if issues:
@@ -242,7 +249,9 @@ def main() -> int:
             print(f"[OK]   {p}")
 
     if total_issues:
-        print(f"\n{len(total_issues)} issue(s) found across {len(md_files)} file(s).", file=sys.stderr)
+        print(
+            f"\n{len(total_issues)} issue(s) found across {len(md_files)} file(s).", file=sys.stderr
+        )
         return 1
     print(f"\nAll {len(md_files)} skill(s) passed validation.")
     return 0
